Move getContentRoot to utils
